+++ index.js
@@ -1,82 +1,8 @@
+/* eslint-env node */
 var fluid = require("infusion");
-var isEqual = require("underscore").isEqual;
+"use strict";
 
-var sjrk = fluid.registerNamespace("sjrk");
 require("./src/couchConfig");
 
-fluid.defaults("sjrk.server.couchConfig.test", {
-    gradeNames: ["sjrk.server.couchConfig.auto"],
-    dbConfig: {
-        dbName: "test",
-        designDocName: "views"
-    },
-    dbDocuments: {
-        "test1": {
-            "message": "Hello, World!",
-            "tags": ["Hello", "World"]
-        },
-        "test2": {
-            "message": "Goodbye, World!",
-            "tags": ["Goodbye", "World"]
-        }
-    },
-    dbViews: {
-        "tags": {
-            "map": "sjrk.server.couchConfig.test.tagsMapFunction"
-        }
-    }
-});
-
-sjrk.server.couchConfig.test.tagsMapFunction = function (doc) {
-    emit("tags", doc.tags);
-};
-
-var testConfig = sjrk.server.couchConfig.test();
-
-fluid.defaults("sjrk.server.couchConfig.stories", {
-    gradeNames: "sjrk.server.couchConfig.auto",
-    dbConfig: {
-        dbName: "stories",
-        designDocName: "views"
-    },
-    dbViews: {
-        titles: {
-            map: "sjrk.server.couchConfig.titleMapFunction"
-        },
-        authors: {
-            map: "sjrk.server.couchConfig.authorMapFunction"
-        },
-        tags: {
-            map: "sjrk.server.couchConfig.tagsMapFunction"
-        },
-        language: {
-            map: "sjrk.server.couchConfig.languageMapFunction"
-        },
-        count: {
-            map: "sjrk.server.couchConfig.countMapFunction",
-            reduce: "_count"
-        }
-    }
-});
-
-sjrk.server.couchConfig.titleMapFunction = function (doc) {
-     emit("title", doc.value.title);
-};
-
-sjrk.server.couchConfig.authorMapFunction = function (doc) {
-     emit("author", doc.value.author);
-};
-
-sjrk.server.couchConfig.tagsMapFunction = function (doc) {
-     emit("tags", doc.value.tags);
-};
-
-sjrk.server.couchConfig.languageMapFunction = function (doc) {
-     emit("language", doc.value.language);
-};
-
-sjrk.server.couchConfig.countMapFunction = function (doc) {
-    emit("count", 1);
-};
-
-// var storiesConfig = sjrk.server.couchConfig.stories();
+// Register our content so that it can be referenced in other packages using `fluid.module.resolvePath("%gpii-binder/path/to/content")`
+fluid.module.register("couch-config", __dirname, require);
